+++ app/workers/service_worker.rb
@@ -1,6 +1,6 @@
 class ServiceWorker < DaemonSpawn::Base
 
-  INTERVAL = 60
+  INTERVAL = 5
 
   WORKER_PID_FILE = Rails.root.join('tmp', 'pids', "service_worker_#{Rails.env}.pid")
   WORKER_LOG_FILE = Rails.root.join('log', "service_worker_#{Rails.env}.log")
